Generation of build artifacts
.\Script\main.py has been renamed to
 .\Script\terminal.py
.\template\2.pyxres has been displaced inside
 .\Script\ and subsenquently terminal.py modified to integrate this change.
 terminal.pyxapp, terminal.html & termina.exe have been generated.

 On a note the pyxel CLI tool doesn't work well with space in name this
  has to be escalated to the pyxel repo.
